corrected redux, changed imports and create logout
