+++ src/components/PrivateRoute.js
@@ -1,9 +1,9 @@
 import { useSelector } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import { authSelectors } from 'redux/auth/authSelectors';
+import { getIsLoggedIn } from 'redux/services/selectors';
 
 export const PrivateRoute = ({ children, redirectTo = '/' }) => {
-  const isLoggedIn = useSelector(authSelectors.getIsLoggedIn);
+  const isLoggedIn = useSelector(getIsLoggedIn);
 
   return !isLoggedIn ? <Navigate to={redirectTo} /> : children;
 };
