+++ src/components/PublicRoute.js
@@ -1,13 +1,13 @@
 import { useSelector } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import { authSelectors } from 'redux/auth/authSelectors';
+import { getIsLoggedIn } from 'redux/services/selectors';
 
 export const PublicRoute = ({
   children,
   redirectTo = '/',
   restricted = false,
 }) => {
-  const isLoggedIn = useSelector(authSelectors.getIsLoggedIn);
+  const isLoggedIn = useSelector(getIsLoggedIn);
 
   const shouldRedirect = isLoggedIn && restricted;
   return shouldRedirect ? <Navigate to={redirectTo} /> : children;
