--- src/redux/products/selectors.js
@@ -1,4 +1,0 @@
-export const selectBadProducts = state => state.products.bad;
-export const selectCalories = state => state.products.calories;
-export const selectLoadStatus = state => state.products.loading;
-export const selectLoadError = state => state.products.error;
