+++ src/redux/services/operations.js
@@ -1,0 +1,63 @@
+import axios from 'axios';
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import toast from 'react-hot-toast';
+
+const BASE_URL = process.env.REACT_APP_BACKEND_URL;
+
+axios.defaults.baseURL = BASE_URL;
+
+const token = {
+  set(token) {
+    axios.defaults.headers.common.Authorization = `Bearer ${token}`;
+  },
+  unset() {
+    axios.defaults.headers.common.Authorization = '';
+  },
+};
+
+export const register = createAsyncThunk(
+  '/api/auth/registration',
+  async credentials => {
+    try {
+      const { data } = await axios.post(`/api/auth/registration`, credentials);
+      token.set(data.token);
+      return data;
+    } catch (error) {}
+  }
+);
+
+export const login = createAsyncThunk(
+  '/api/auth/login',
+  async (data, thunkAPI) => {
+    try {
+      const res = await await axios.post('/api/auth/login', data);
+      return res.data;
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error.message);
+    }
+  }
+);
+
+export const getProducts = createAsyncThunk(
+  '/products/getBad',
+  async (userParams, thunkAPI) => {
+    try {
+      const { data, status } = await axios.post('/products', userParams);
+      token.set(data.token);
+      if (!data) {
+        return thunkAPI.rejectWithValue(status);
+      }
+      return data;
+    } catch (err) {
+      toast('Get products error');
+      return thunkAPI.rejectWithValue(err.response.data);
+    }
+  }
+);
+
+ export const logout = createAsyncThunk('/api/auth/logout', async () => {
+   try {
+     await axios.get(`/api/auth/logout`);
+     token.unset();
+   } catch (error) {}
+ });
