--- src/redux/store.js
@@ -1,10 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { authReducer } from './Slice/loginSlice';
-import { productsReducer } from './products/productSlice';
-
-export const store = configureStore({
-  reducer: {
-    auth: authReducer,
-    products: productsReducer,
-  },
-});
