+++ src/redux/store/store.js
@@ -11,8 +11,8 @@ import {
   PURGE,
   REGISTER,
 } from 'redux-persist';
-import { authReducer } from '../auth/authSlice';
-import { productsReducer } from '../products/productSlice';
+import { authReducer } from '../services/authSlice';
+import { productsReducer } from '../services/productSlice';
 
 const persistAuthConfig = {
   key: 'auth',
